+++ CHANGELOG.md
@@ -12,11 +12,12 @@ changes.
 
 ### Added
 
--
+- Unsancho GovTool [Issue 1637](https://github.com/IntersectMBO/govtool/issues/1637)
+- Add network metrics model to frontend service
 
 ### Fixed
 
--
+- Incorrect copy on DRep votes + other minor copy spelling mistakes
 
 ### Changed